historial

+++ database/migrations/2025_11_07_041855_create_materias_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('nombre', 150);
             $table->integer('id_tipo_competencia')->nullable()->index('id_tipo_competencia');
             $table->integer('id_modalidad')->nullable()->index('id_modalidad');
-            $table->integer('creditos')->nullable();
+            $table->decimal('creditos', 5)->nullable();
             $table->integer('horas')->nullable();
             $table->integer('id_espacio_formativo')->nullable()->index('id_espacio_formativo');
             $table->integer('id_plan_estudio')->nullable()->index('id_plan_estudio');
